an approach Shannon-Fano, probably wrong
I think I will need to completely redo this part, the Shannon-Fano construction
is going to need to rely on there being exactly one Node (probably a special
kind of node) starting off, we can render the consistuent codes inside that but
they don't exist as separate entities. Then we split the node (this should be an
option in the dropdown), now we have two active nodes that we can drag and drop
between, initially we'll start everything in one of them.

We can use a click to active the split mode, but this may be too easy to trigger
accidentally (so maybe there should be an "abort split" option in the dropdown).